requirements.txt
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
MarkupSafe==3.0.3
python-dotenv==1.1.1
Werkzeug==3.1.3
tensorflow==2.20.0
numpy==2.3.4